requirements_production.txt
# Core dependencies (optimized versions)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
psutil==5.9.8

# Email processing (minimal)
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Vector search and embeddings (optimized)
numpy==1.24.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.19.4
nomic==2.0.0

# LLM providers (only what you need)
cohere==4.37
groq==0.5.0
google-generativeai==0.3.2
openai==1.3.0

# Utilities (minimal)
pandas==2.0.3
tqdm==4.66.1
python-dateutil==2.8.2

# Production dependencies
gunicorn==21.2.0 